Update create.jinja2
Good thought process here about how the type of a form input might relate to the Python type that you're going to want to use for the variable.

Unfortunately, HTML forms and Python aren't able to communicate this closely with each other. The closest we can do is `type=number`.